Move status scheduler to status file
