+++ commands/maps/status.js
@@ -10,7 +10,13 @@ const {
   codeBlock,
 } = require('../../helpers');
 const { v4: uuidv4 } = require('uuid');
-const { insertNewStatus, getStatus, deleteStatus } = require('../../database/handler');
+const {
+  insertNewStatus,
+  getStatus,
+  deleteStatus,
+  getAllStatus,
+} = require('../../database/handler');
+const Scheduler = require('../../scheduler');
 
 //----- Status Application Command Replies -----//
 /**
@@ -360,6 +366,58 @@ const cancelStatusStop = async ({ nessie, interaction }) => {
     await sendErrorLog({ nessie, error, interaction, type, uuid });
   }
 };
+const initialiseStatusScheduler = (nessie) => {
+  return new Scheduler('10 */5 * * * *', async () => {
+    getAllStatus(
+      async (allStatus) => {
+        try {
+          const rotationData = await getRotationData();
+          const statusLogChannel = nessie.channels.cache.get('976863441526595644');
+          if (allStatus) {
+            allStatus.forEach(async (status) => {
+              const pubsChannel = nessie.channels.cache.get(status.pubs_channel_id);
+              const rankedChannel = nessie.channels.cache.get(status.ranked_channel_id);
+              const pubsMessage = await pubsChannel.messages.fetch(status.pubs_message_id);
+              const rankedMessage = await rankedChannel.messages.fetch(status.ranked_message_id);
+
+              const pubsEmbed = generatePubsStatusEmbeds(rotationData);
+              const rankedEmbed = generateRankedStatusEmbeds(rotationData);
+
+              await pubsMessage.edit({ embeds: pubsEmbed });
+              await rankedMessage.edit({ embeds: rankedEmbed });
+              //Figure out rate limiting prevention here
+              //Docs say normal requests is 50 per second but idk if this falls into a special route case
+              //Probably just take a risk and try to send 40 requests (10 servers) and then add a timeout of 1 second?
+            });
+          }
+          const statusLogEmbed = {
+            title: 'Nessie | Auto Map Status Log',
+            description: 'Requested data from API and checked database',
+            timestamp: Date.now(),
+            color: 3066993,
+            fields: [
+              {
+                name: 'Auto Map Status Count:',
+                value: allStatus ? `${allStatus.length}` : '0',
+                inline: true,
+              },
+            ],
+          };
+          await statusLogChannel.send({ embeds: [statusLogEmbed] });
+        } catch (error) {
+          const uuid = uuidv4();
+          const type = 'Status Scheduler (Editing)';
+          await sendErrorLog({ nessie, error, type, uuid, ping: true });
+        }
+      },
+      async (error) => {
+        const uuid = uuidv4();
+        const type = 'Status Scheduler (Database)';
+        await sendErrorLog({ nessie, error, type, uuid, ping: true });
+      }
+    );
+  });
+};
 module.exports = {
   /**
    * Creates Status application command with relevant subcommands
@@ -410,4 +468,5 @@ module.exports = {
   deleteStatusChannels,
   generatePubsStatusEmbeds,
   generateRankedStatusEmbeds,
+  initialiseStatusScheduler,
 };
