+++ events.js
@@ -34,6 +34,7 @@ const {
   deleteStatusChannels,
   generatePubsStatusEmbeds,
   generateRankedStatusEmbeds,
+  initialiseStatusScheduler,
 } = require('./commands/maps/status');
 const Scheduler = require('./scheduler');
 const { v4: uuidv4 } = require('uuid');
@@ -66,60 +67,8 @@ exports.registerEventHandlers = ({ nessie, mixpanel }) => {
       nessie.user.setActivity(brPubsData.current.map); //Set current br map as activity status
       sendHealthLog(brPubsData, logChannel, true); //For logging purpose
       setCurrentMapStatus(brPubsData, logChannel, nessie); //Calls status display function
-
-      const statusScheduler = new Scheduler('10 */5 * * * *', async () => {
-        getAllStatus(
-          async (allStatus) => {
-            try {
-              const rotationData = await getRotationData();
-              const statusLogChannel = nessie.channels.cache.get('976863441526595644');
-              if (allStatus) {
-                allStatus.forEach(async (status) => {
-                  const pubsChannel = nessie.channels.cache.get(status.pubs_channel_id);
-                  const rankedChannel = nessie.channels.cache.get(status.ranked_channel_id);
-                  const pubsMessage = await pubsChannel.messages.fetch(status.pubs_message_id);
-                  const rankedMessage = await rankedChannel.messages.fetch(
-                    status.ranked_message_id
-                  );
-
-                  const pubsEmbed = generatePubsStatusEmbeds(rotationData);
-                  const rankedEmbed = generateRankedStatusEmbeds(rotationData);
-
-                  await pubsMessage.edit({ embeds: pubsEmbed });
-                  await rankedMessage.edit({ embeds: rankedEmbed });
-                  //Figure out rate limiting prevention here
-                  //Docs say normal requests is 50 per second but idk if this falls into a special route case
-                  //Probably just take a risk and try to send 40 requests (10 servers) and then add a timeout of 1 second?
-                });
-              }
-              const statusLogEmbed = {
-                title: 'Nessie | Auto Map Status Log',
-                description: 'Requested data from API and checked database',
-                timestamp: Date.now(),
-                color: 3066993,
-                fields: [
-                  {
-                    name: 'Auto Map Status Count:',
-                    value: allStatus ? `${allStatus.length}` : '0',
-                    inline: true,
-                  },
-                ],
-              };
-              await statusLogChannel.send({ embeds: [statusLogEmbed] });
-            } catch (error) {
-              const uuid = uuidv4();
-              const type = 'Status Scheduler (Editing)';
-              await sendErrorLog({ nessie, error, type, uuid, ping: true });
-            }
-          },
-          async (error) => {
-            const uuid = uuidv4();
-            const type = 'Status Scheduler (Database)';
-            await sendErrorLog({ nessie, error, type, uuid, ping: true });
-          }
-        );
-      });
-      statusScheduler.start();
+      const statusScheduler = initialiseStatusScheduler(nessie); //Initialises auto status scheduler
+      statusScheduler.start(); //Starts the status scheduler
     } catch (e) {
       console.log(e); //Add proper error handling
     }
